--- examples/linebot/src/admin.ts
@@ -1,12 +1,0 @@
-
-
-import Client, { Player } from '../../../dist/index.js'
-import client from './line.js'
-
-export function is_bot_admin(player: Player) {
-    return player.cuid == client.self?.cuid
-}
-
-export function module (client: Client) {
-    return client
-}
